requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
python-multipart
jinja2